% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\alias{MLE}
\alias{maximumlikelihoodestimator}
\title{maximum likelihood estimator}
\usage{
mle(y)
}
\arguments{
\item{y}{The vector of observed data to calculate the log likelihood from}
}
\value{
A single numeric output:
 \item{MLE}{The maximum likelihood estimate of lamda}
}
\description{
Calculates the maximum likelihood estimator for lambda in a poisson distribution
}
\note{
This uses the function for a poisson distribution. Do not use with data from another distribution.
}
\examples{

set.seed(1625)
y <- rpois(1000,80)
mle(y)

}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{standardError}}
}
\author{
Jordon Newton
}
