% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMLE.R
\name{plotMLE}
\alias{plotMLE}
\alias{plotmle}
\title{Plot MLE}
\usage{
plotMLE(y, SEtype, B = 1000, lambda_width = 2.5)
}
\arguments{
\item{y}{The vector of observed data}

\item{SEtype}{The type of SE to be calculated; basic, or bootstrapped.}

\item{B}{The number of bootstraps to run, default set to 1000 (not required for basic SEs). Default is 1000}

\item{lambda_width}{The distance from the estimated MLE to be included in the plot on either side of the MLE. Default is 2.5.}
}
\value{
A plot
 \item{MLE Plot}{A plot that shows potential values of lambda, and the log likelihood of each, with the standard errors of MLE surrounding the maximum estimate.}
}
\description{
Plots the MLE with standard errors.
}
\note{
The MLE used is calculated using the mle() function from this package, the log likelihood using logLik, and the SEs using "standardError".
}
\examples{

y <- rpois(1000,80)
plotMLE(y,"basic")
plotMLE(y,"bootstrapped",100,5)

}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{mle}}

\code{\link{standardError}}
}
\author{
Jordon Newton
}
