% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\alias{SE}
\title{Standard Error}
\usage{
standardError(y, SEtype, B = 1000)
}
\arguments{
\item{y}{The vector of observed data}

\item{SEtype}{The type of SE to be calculated; basic, or bootstrapped.}

\item{B}{The number of bootstraps to run, default set to 1000. Not used for basic SE estimation.}
}
\value{
A single numeric output:
 \item{SE}{Standard error from the type that was chosen, "basic", or "bootstrapped".}
}
\description{
Calculates the standard error for the data, using one or two methods; standard, or bootstrap
}
\note{
The MLE used is calculated using the mle() function from this package. The basic SE is the square root of MLE/n. The bootstrapped version samples n observations from the dataset with replacement B times. The resulting SE is the SD of the MLEs from this matrix.
}
\examples{

set.seed(1625)
y <- rpois(1000,80)
standardError(y,"basic")
standardError(y,"bootstrapped",1000)

}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{mle}}
}
\author{
Jordon Newton
}
