% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{maximum likelihood estimator}
\usage{
mle(y)
}
\arguments{
\item{y}{The vector of observed data to calculate the log likelihood from}
}
\value{
A single numeric output:
 \item{MLE}{The log maximum likelihood estimate of lamda}
}
\description{
Calculates the maximum likelihood estimator for lambda in a poisson distribution
}
\note{
This uses the basic formula to estimate the MLE
}
\examples{



}
\author{
Jordon Newton
}
