% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\title{Standard Error}
\usage{
standardError(y, SEtype, B)
}
\arguments{
\item{y}{The vector of observed data}

\item{SEtype}{The type of SE to be calculated; basic, or bootstrapped.}

\item{B}{The number of bootstraps to run, default set to 1000 (not required for basic SEs).}
}
\value{
A single numeric output:
 \item{SE}{Standard error from the type that was chosen, "basic", or "bootstrapped".}
}
\description{
Calculates the standard error for the data, using one or two methods; standard, or bootstrap
}
\note{
The basic SE is the square root of MLE/n. The bootstrapped version samples n observations from the dataset with replacement B times. The resulting SE is the SD of the MLEs from this matrix.
}
\examples{



}
\author{
Jordon Newton
}
